feat: Change upvote to toggleUpvote
