+++ lambda_backend/route_microservice/serverless.yml
@@ -107,11 +107,11 @@ functions:
           cors:
             origin: ${self:custom._origin.${self:custom.stage}}
 
-  upvoteRoute:
-    handler: src/upvoteRoute.handler
+  toggleUpvoteRoute:
+    handler: src/toggleUpvoteRoute.handler
     events:
       - http:
-          path: route/details/upvote
+          path: route/details/toggleUpvote
           method: post
           cors:
             origin: ${self:custom._origin.${self:custom.stage}}
